Add Genome Driven Diagnosis on Patient View
Uses an internal service to get the cancer type prediction for each impact
case. This cancer type prediction is called Genome Driven Diagnosis and is a
predicted probability for 22 tumor types based on genomic features, using a
calibrated RandomForest algorithm. We would like to indicate mismatches between
these and pathology based assessments of cancer type. Furthermore it's useful
for samples of unkown primary. Currently the service will only be available
internally at MSKCC.

The GDD predicted cancer type shows up in the patient view header when it
differs from the diagnosed cancer type. In red if it's a high confident
mismatch (>75%), in orange if it's a medium confident mismatch (>50%),
otherwise it shows up in blue. The top bar in the tooltip is colored similarly
(just gray instead of blue for low confident calls). When the cancer type
matches the diagnosed one the high confident match is bar is colored green, the
medium confident match is blue. The coloring thresholds are the same as
@anoopbr used for CVR.

When the diagnosed cancer type is not part of the 22 cancer types a warning
message is displayed in the tooltip. Similarly when the diagnosed cancer type
is undefined.